refactor: change package directory structure and implement minor changes
Add status and error return to MaybeHandshake
Change package directory to be more modular
Mark future TODO for handshakeHandler
