+++ internal/connection/connection.go
@@ -2,10 +2,6 @@ package connection
 
 import "github.com/khorcarol/AgentOfThings/internal/api"
 
-type Config struct {
-	X int
-}
-
 var (
 	NewUserChannel       chan api.User   = make(chan api.User)
 	FriendRequestChannel chan api.Friend = make(chan api.Friend)
