+++ internal/connection/protocol/handshake/peer_to_user/handshake_test.go
@@ -1,4 +1,4 @@
-package peer_to_user_handshake
+package peer_to_user
 
 import (
 	"bytes"
